Fix incorrect property keys when importing GDS files
This pull request addresses an issue where property keys are incorrectly interpreted as signed integers when importing GDS files using the `import_gds` function in `phidl.geometry`. The issue is caused by a bug in the `gdspy` library (heitzmann/gdspy#240).

When saving properties using `gdspy`, the keys are directly written to the GDS file as unsigned 16-bit integers (2 bytes). However, when loading the GDS file, `gdspy` interprets the keys as signed integers, resulting in negative values for keys greater than 32767.

This leads to incorrect property keys in the loaded GDS elements when using `phidl.geometry.import_gds`.

To work around this issue, a new function `_correct_properties` has been added to `phidl.geometry`. This function corrects the properties dictionary of a loaded GDS element by converting the keys from the signed integer range to the original unsigned integer range.

The `import_gds` function has been modified to call `_correct_properties` on the properties of each loaded GDS element (`DeviceReference` and `CellArray`) before assigning them to the corresponding `phidl` objects.

The `_correct_properties` function modifies the input `properties` dictionary in place. For each key-value pair, if the key is negative, it adds 65536 (2**16) to the key to convert it back to the original unsigned integer value. The value is then stored under the new key, and the negative key is marked for deletion. Finally, the marked keys are deleted from the dictionary.

This fix ensures that the property keys are correctly imported when loading GDS files using `phidl.geometry.import_gds`. Users of `phidl` who rely on the `import_gds` function to import GDS files with properties will benefit from this fix, as it will prevent incorrect property keys in the loaded GDS elements.

- heitzmann/gdspy#240

---